requirements.txt
drf-yasg==1.20.0
django==4.0.4
djangorestframework==3.13.1
vn
djoser

